requirements.txt
fastapi==0.109.0          # Web framework
uvicorn[standard]==0.27.0 # ASGI server with extras
httpx==0.26.0            # Async HTTP client
pydantic==2.5.0          # Data validation
pydantic-settings==2.1.0 # Settings management
python-dotenv==1.0.0     # Environment variables
redis==5.0.1             # Caching (optional)
celery==5.3.4            # Background tasks (optional)
python-json-logger==2.0.7 # Structured logging
tenacity==8.2.3          # Retry logic
pytest==7.4.3            # Testing
pytest-asyncio==0.23.0   # Async test support
black==23.12.0           # Code formatting
ruff==0.1.9              # Linting